next step

+++ deeplearning4j-scaleout/spark/dl4j-spark/src/test/java/org/deeplearning4j/spark/datavec/TestPreProcessedData.java
@@ -262,7 +262,7 @@ public class TestPreProcessedData extends BaseSparkTest {
 
     private static String toString(INDArray rowVector, int labelIdx) {
         StringBuilder sb = new StringBuilder();
-        int length = rowVector.length();
+        long length = rowVector.length();
         for (int i = 0; i < length; i++) {
             sb.append(rowVector.getDouble(i));
             sb.append(",");
